+++ src/components/ServiceHeader/ServiceHeader.module.css
@@ -70,6 +70,7 @@
     border-radius: 8px;
     background-color: var(--white);
     box-shadow: 0 2px 12px rgba(0, 0, 0, 0.08);
+    z-index: 100;
 
     > button {
       width: 130px;
