+++ src/pages/List.jsx
@@ -7,9 +7,8 @@ import "swiper/css/pagination"
 import "swiper/css/scrollbar"
 import styles from "./List.module.css"
 import useRequest from "../hooks/useRequest"
-import CardList from '../components/CardList/CardList'
+import CardList from "../components/CardList/CardList"
 import ButtonPrimary from "../components/Button/ButtonPrimary"
-
 
 const RECIPIENTS_LIMIT = 100
 
@@ -86,9 +85,7 @@ const List = () => {
         </div>
       </secion>
       <secion className={styles.button}>
-        <ButtonPrimary link="/post">
-          나도 만들어보기
-        </ButtonPrimary>
+        <ButtonPrimary link="/post">나도 만들어보기</ButtonPrimary>
       </secion>
     </div>
   )
