+++ src/pages/PostId_PostIdEdit/PostId.jsx
@@ -1,10 +1,10 @@
 import React, { useCallback, useEffect, useState } from "react"
 import { useNavigate, useParams } from "react-router-dom"
 import ServiceHeader from "../../components/ServiceHeader/ServiceHeader"
-import Card from "../../components/Card"
-import CardAdd from "../../components/CardAdd"
-import Modal from "../../components/Modal"
-import ButtonPrimary from "../../components/ButtonPrimary"
+import Card from "../../components/Card/Card"
+import CardAdd from "../../components/CardAdd/CardAdd"
+import Modal from "../../components/Modal/Modal"
+import ButtonPrimary from "../../components/Button/ButtonPrimary"
 import useRequest from "../../hooks/useRequest"
 import styles from "./PostId.module.css"
