+++ src/pages/PostId_PostIdEdit/PostId.module.css
@@ -68,7 +68,6 @@
   right: 0px;
 }
 
-
 @media (min-width: 375px) and (max-width: 1199px) {
   .delete-page-button {
     height: 55px;
@@ -78,8 +77,9 @@
     position: fixed;
     bottom: 24px;
     top: auto;
-    right:auto;
+    right: auto;
     left: auto;
+    z-index: 1;
   }
 }
 
@@ -90,7 +90,7 @@
   .card-list-container {
     max-width: 720px;
   }
-  .delete-page-button{
+  .delete-page-button {
     width: 720px;
   }
 }
@@ -101,7 +101,7 @@
   .card-list-container {
     max-width: 352px;
   }
-  .delete-page-button{
+  .delete-page-button {
     width: 352px;
   }
 }
