+++ src/utils/stripHtml.js
@@ -1,0 +1,3 @@
+export function stripHtml(html) {
+  return html.replace(/<[^>]*>?/gm, '');
+}
